Move AddSendOnlyNServiceBusEndpoint
