+++ src/Tests/ApprovalFiles/ApiApprovals.ApproveAzureServiceBusComponentApi.approved.txt
@@ -18,11 +18,6 @@ namespace NServiceBus
         public override System.Collections.Generic.IReadOnlyCollection<NServiceBus.TransportTransactionMode> GetSupportedTransactionModes() { }
         public override System.Threading.Tasks.Task<NServiceBus.Transport.TransportInfrastructure> Initialize(NServiceBus.Transport.HostSettings hostSettings, NServiceBus.Transport.ReceiveSettings[] receivers, string[] sendingAddresses, System.Threading.CancellationToken cancellationToken = default) { }
     }
-    public static class FunctionsHostApplicationBuilderExtensions
-    {
-        public static void AddSendOnlyNServiceBusEndpoint(this Microsoft.Azure.Functions.Worker.Builder.FunctionsApplicationBuilder builder, string endpointName, System.Action<NServiceBus.EndpointConfiguration> configure) { }
-        public static void AddSendOnlyNServiceBusEndpoint(this Microsoft.Azure.Functions.Worker.Builder.FunctionsApplicationBuilder builder, string endpointName, System.Action<NServiceBus.EndpointConfiguration, Microsoft.Extensions.DependencyInjection.IServiceCollection> configure) { }
-    }
 }
 namespace NServiceBus.Configuration.AdvancedExtensibility
 {
